Moved config
