+++ packages/mds/stencil.config.docs.ts
@@ -1,6 +1,5 @@
 import { Config } from '@stencil/core';
-import { docsWww } from './config/outputTargets';
-import { config as baseConfig } from './stencil.config'
+import { config as baseConfig, docsWww } from './stencil.config'
 
 // Only builds the docs output.
 export const config: Config = {
