+++ packages/mds/stencil.config.mds.ts
@@ -1,6 +1,5 @@
 import { Config } from '@stencil/core';
-import { dist, distCustomElements, hydrate } from './config/outputTargets';
-import { config as baseConfig } from './stencil.config'
+import { config as baseConfig, dist, distCustomElements, hydrate } from './stencil.config'
 
 // Only builds the outputs required for the mds package export
 export const config: Config = {
