+++ packages/mds/stencil.config.react.ts
@@ -1,6 +1,5 @@
 import { Config } from '@stencil/core';
-import { dist, react } from './config/outputTargets';
-import { config as baseConfig } from './stencil.config'
+import { config as baseConfig, dist, react } from './stencil.config'
 
 // Only builds the outputs requireed for react
 // React needs the dist for the loader export
